refact: split app into small files
