+++ src/config/index.ts
@@ -1,0 +1,1 @@
+export {default as theme} from './theme'
