+++ src/config/theme.ts
@@ -1,0 +1,4 @@
+export default {
+  useSystemColorMode: false,
+  initialColorMode: "dark",
+}
